refactor(express): support request handler for serverless functions
