+++ packages-backend/express/src/index.ts
@@ -1,2 +1,3 @@
-export * from './constants';
+export { CLOUD_IDENTIFIERS, MC_API_URLS } from './constants';
+export { createSessionAuthVerifier } from './auth';
 export { default as createSessionMiddleware } from './middlewares/session-middleware';
