+++ packages-backend/express/src/types.ts
@@ -1,4 +1,3 @@
-import type { Request } from 'express';
 import type { ExpressJwtOptions } from 'jwks-rsa';
 
 import { CLOUD_IDENTIFIERS } from './constants';
@@ -6,6 +5,10 @@ import { CLOUD_IDENTIFIERS } from './constants';
 export type TCloudIdentifier = typeof CLOUD_IDENTIFIERS[keyof typeof CLOUD_IDENTIFIERS];
 
 export type TSessionMiddlewareOptions = {
+  // The public-facing URL used to connect to the server / serverless function.
+  // The value should only contain the origin URL (protocol, hostname, port),
+  // the request path is inferred from the incoming request.
+  audience: string;
   // The cloud identifier (see `CLOUD_IDENTIFIERS`) that maps to the MC API URL
   // of the related cloud region or the MC API URL.
   issuer: TCloudIdentifier | string;
@@ -23,5 +26,3 @@ export type TSession = {
   userId: string;
   projectKey: string;
 };
-
-export type TRequestWithSession = Request & { session: TSession };
